Custom Qt SPICE viewer
a basic Qt SPICE viewer. This allows users to interact with a VM through input (mouse, keyboard, etc...) as well as see its display and receive audio. As planned, this is **rendered onto a KDE-native window** and **virt-viewer is no longer required as a dependency**.

Note: Previous VMs configured in Karton will not have audio. You can either add a device in the XML file for create a new one with this branch.